% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brazil_state}
\alias{brazil_state}
\title{The dataset with information on the new Coronavirus (COVID-19) for the twenty-seven states of Brazil}
\format{
A data.frame object
}
\source{
Ministerio da Saude - Sistema Unico de Saude (SUS) \href{https://www.gov.br/saude/pt-br}{website}
}
\usage{
brazil_state
}
\value{
load the database for the current environment
}
\description{
Daily information about the new Coronavirus (COVID-19) pandemic in Brazil's states
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Brazil's states. The data includes the following fields:

* region - String. Indicates to which region the state belongs

* state - String. Acronym (abbreviation) for state name

* coduf - Numeric. Number that identifies the state

* date - Date. Ocurrence notification date

* epidWeek - Numeric. Epidemiological week of ocurrence by notification date

* population - Numeric. Population of the size, of each state, on the date of notification of the ocurrence

* accumCases - Numeric. Cumulative cases of COVID-19 by notification date

* newCases - Numeric. Number of new COVID-19 cases by notification date

* accumDeaths - Numeric. Accumulated COVID-19 deaths by notification date

* newDeaths - Numeric. Number of new COVID-19 deaths by notification date
}
\examples{

data (brazil_state)

}
\keyword{Brazil}
\keyword{COVID-19}
\keyword{coronavirus}
\keyword{datasets}
\keyword{sars-cov-2}
