% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brazil_total}
\alias{brazil_total}
\title{The dataset with information about the new Coronavirus (COVID-19) for Brazil}
\format{
A data.frame object
}
\source{
Ministerio da Saude - Sistema Unico de Saude (SUS) \href{https://www.gov.br/saude/pt-br}{website}
}
\usage{
brazil_total
}
\value{
load the database for the current environment
}
\description{
Daily information about the new Coronavirus (COVID-19) pandemic in Brazil
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Brazil. The data includes the following fields:

* region - String. Indicates the country of coverage

* date - Date. Ocurrence notification date

* epidWeek - Numeric. Epidemiological week of ocurrence by notification date

* population - Numeric. Population size at the date of ocurrence notification

* accumCases - Numeric. Cumulative cases of COVID-19 by notification date

* newCases - Numeric. Number of new COVID-19 cases by notification date

* accumDeaths - Numeric. Accumulated COVID-19 deaths by notification date

* newDeaths - Numeric. Number of new COVID-19 deaths by notification date

* newRecov - Numeric. Number of new recovered COVID-19 patients by notification date

* followUp - Numeric. Number of new COVID-19 cases being followed up by notification date
}
\examples{

data (brazil_total)

}
\keyword{Brazil}
\keyword{COVID-19}
\keyword{coronavirus}
\keyword{datasets}
\keyword{sars-cov-2}
