% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brazil_municipality}
\alias{brazil_municipality}
\title{The dataset with information on the new Coronavirus (COVID-19) for the 5,570 municipalities in Brasil}
\format{
A data.frame object
}
\source{
Ministerio da Saude - Sistema Unico de Saude (SUS) \href{https://www.saude.gov.br/}{website}
}
\usage{
brazil_municipality
}
\description{
Daily information about the new Coronavirus (COVID-19) pandemic cases in Brazil's municipalities
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Brazil's municipalities. The data includes the following fields:

- region - indicates to which region the municipality belongs

- state - indicates to which state the municipality belongs

- municipality - municipality name

- coduf - number that identifies the state

- codmun - number that identifies the municipality

- date - date of notification

- epidWeek - epidemiological week

- population - population size

- accumCases - number of cases accumulated

- newCases - number of new cases reported

- accumDeaths - number of accumulated deaths

- newDeaths - new deaths reported
}
\examples{

data (brazil_municipality)

head (brazil_municipality)

}
\keyword{COVID19}
\keyword{Italy}
\keyword{coronavirus}
\keyword{datasets}
