% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brazil_region}
\alias{brazil_region}
\title{The dataset with information on the new Coronavirus (COVID-19) for the five regions of Brazil}
\format{
A data.frame object
}
\source{
Ministério da Saúde - Sistema Único de Saúde (SUS) \href{https://www.saude.gov.br/}{website}
}
\usage{
brazil_region
}
\description{
Daily information about the new Coronavirus (COVID-19) pandemic in Brazil's regions
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Brazil's regions. The data includes the following fields:

- region - indicates the regions coverage (Centro-Oeste, Nordeste, Norte, Sudeste, Sul)

- date - date of notification

- epidWeek - epidemiological week

- population - population size

- accumCases - number of cases accumulated

- newCases - number of new cases reported

- accumDeaths - number of accumulated deaths

- newDeaths - new deaths reported

- geom - georeferencing polygons (multipolygon) for each region
}
\examples{

data (brazil_region)

head (brazil_region)

}
\keyword{Brazil}
\keyword{COVID-19}
\keyword{coronavirus}
\keyword{datasets}
\keyword{sars-cov-2}
