% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brazil_state}
\alias{brazil_state}
\title{The dataset with information on the new Coronavirus (COVID-19) for the twenty-seven states of Brazil}
\format{
A data.frame object
}
\source{
Ministério da Saúde - Sistema Único de Saúde (SUS) \href{https://www.saude.gov.br/}{website}
}
\usage{
brazil_state
}
\description{
Daily information about the new Coronavirus (COVID-19) pandemic in Brazil's states
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Brazil's states. The data includes the following fields:

- region - indicates to which region the state belongs

- state - acronym (abbreviation) for state name

- coduf - number that identifies the state

- date - date of notification

- epidWeek - epidemiological week

- population - population size

- accumCases - number of cases accumulated

- newCases - number of new cases reported

- accumDeaths - number of accumulated deaths

- newDeaths - new deaths reported

- newRecov - indicates the number of new recovered patients

- followUp - indicates the number of new cases being followed up

- geom - georeferencing polygons (multipolygon) for each state
}
\examples{

data(brazil_state)

head(brazil_state)

}
\keyword{COVID19}
\keyword{Italy}
\keyword{coronavirus}
\keyword{datasets}
