% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brazil_total}
\alias{brazil_total}
\title{The dataset with information about the new Coronavirus (COVID-19) for Brazil}
\format{
A data.frame object
}
\source{
Ministerio da Saude - Sistema Unico de Saude (SUS) \href{https://www.saude.gov.br/}{website}
}
\usage{
brazil_total
}
\description{
Daily information about the new Coronavirus (COVID-19) pandemic in Brazil
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Brazil. The data includes the following fields:

- region - indicates the country of coverage

- date - date of notification

- epidWeek - epidemiological week

- population - population size

- accumCases - number of cases accumulated

- newCases - number of new cases reported

- accumDeaths - number of accumulated deaths

- newDeaths - new deaths reported

- newRecov - indicates the number of new recovered patients

- followUp - indicates the number of new cases being followed up
}
\examples{

data (brazil_total)

head (brazil_total)

}
\keyword{Brazil}
\keyword{COVID-19}
\keyword{coronavirus}
\keyword{datasets}
\keyword{sars-cov-2}
